+++ cmd/vega/nodewallet/show.go
@@ -7,6 +7,7 @@ import (
 	"code.vegaprotocol.io/vega/config"
 	"code.vegaprotocol.io/vega/logging"
 	"code.vegaprotocol.io/vega/nodewallets"
+	"code.vegaprotocol.io/vega/nodewallets/registry"
 
 	"github.com/jessevdk/go-flags"
 )
@@ -37,12 +38,12 @@ func (opts *showCmd) Execute(_ []string) error {
 		return err
 	}
 
-	registryLoader, err := nodewallets.NewRegistryLoader(vegaPaths, registryPass)
+	registryLoader, err := registry.NewLoader(vegaPaths, registryPass)
 	if err != nil {
 		return err
 	}
 
-	registry, err := registryLoader.GetRegistry(registryPass)
+	registry, err := registryLoader.Get(registryPass)
 	if err != nil {
 		return err
 	}